Variante v0.3.2
Klassennamen Kunde und Rechnung korrigiert (Großgeschrieben am Anfang)
